requirements.txt
streamlit
openai
langchain
langchain_openai
pinecone-client
langchain_community
pinecone-text
huggingface-hub
rank_bm25
langchain-pinecone
nltk>=3.6.0